+++ src/main/java/me/draconix6/moveresizeplugin/command/CursorSpeedCommand.java
@@ -1,19 +1,11 @@
-package me.draconix6.moveresizeplugin;
+package me.draconix6.moveresizeplugin.command;
 
-import com.sun.jna.Pointer;
-import com.sun.jna.platform.win32.WinDef;
+import me.draconix6.moveresizeplugin.MoveResizePlugin;
 import org.apache.logging.log4j.Level;
 import xyz.duncanruns.julti.Julti;
-import xyz.duncanruns.julti.JultiOptions;
 import xyz.duncanruns.julti.cancelrequester.CancelRequester;
 import xyz.duncanruns.julti.command.Command;
-import xyz.duncanruns.julti.command.CommandFailedException;
-import xyz.duncanruns.julti.instance.MinecraftInstance;
-import xyz.duncanruns.julti.management.InstanceManager;
-import xyz.duncanruns.julti.util.WindowStateUtil;
-import win32.User32;
-
-import java.awt.*;
+import me.draconix6.moveresizeplugin.win32.User32;
 
 public class CursorSpeedCommand extends Command {
