+++ src/main/java/me/draconix6/moveresizeplugin/win32/HwndUtil.java
@@ -1,4 +1,4 @@
-package win32;
+package me.draconix6.moveresizeplugin.win32;
 
 import com.sun.jna.Pointer;
 import com.sun.jna.platform.win32.Kernel32;
