+++ src/main/java/me/draconix6/moveresizeplugin/win32/User32.java
@@ -1,4 +1,4 @@
-package win32;
+package me.draconix6.moveresizeplugin.win32;
 
 import com.sun.jna.Native;
 import com.sun.jna.Pointer;
