+++ src/main/java/xyz/duncanruns/eyesee/OverlayGUI.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="eyesee.OverlayGUI">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="xyz.duncanruns.eyesee.OverlayGUI">
   <grid id="27dc6" binding="panel1" layout-manager="BorderLayout" hgap="0" vgap="0">
     <constraints>
       <xy x="15" y="18" width="138" height="68"/>
